refactor: split TTS service per platform
